Add 'Open in Editor' function
This adds a context menu entry to open the selected
molecule in a molecular editor (currently Avogadro).

This uses chemkit to predict and optimize 3D coordinates
for the molecule created from its InChI formula. The
molecule is then written to a temporary CML file which
is passed on the command line to Avogadro.

Change-Id: I1d935ed7de1b02b005a5aa8e8af151ac5467bc45