Fix bug with context menu and substructure queries
This fixes a bug in which the 'Open in Editor' function in
the MongoTableView context-menu would not be displayed when
looking at the results of a substructure query.

The issue was that the BSONObj* stored inside the MongoModel
was not being passed up through the QSortFilterModel used
for substructure queries. The fix is to map the index from
the filter model to the source model when building to menu.

Change-Id: I276470fd1f41ecab5e7c5cd89b7885ee6da8bb87