Add support for similarity clustering
This adds two new dialogs which display a graph showing the
similarity between a group of molecules.

The first is the FingerprintSimilarityDialog which calculates
the similarity of the molecules based on the similarity of
their molecular fingerprints.

The second is the StructureSimilarityDialog which calculates
the similarity of the molecules based on the similarity
between their molecular structures.

Change-Id: Ie93ea059fa3e27c561b710fefb3201b33ca117dc