Add support for tag auto-completion
This adds support for auto-completing a tag when the user is
adding a new tag for a molecule.

For now, the entire molecules collection is queried and a set
of all tags is built which is then passed to the QCompleter
object. In the future this could be made more efficient by storing
a document in mongo with just the set of seen tags and only querying
that when the user starts typing.

Change-Id: I8f659f1f69e7c674300534b949c25207cc71f51e