Add Import CSV File dialog
This adds a new dialog which allows the user to import molecular
descriptor data from CSV files. For now, the dialog expects a CSV
file containing the InChIKey as the first column followed by one
or more columns with the descriptor values.

Also adds a small script for generating CSV files with molecular
descriptor values for a molecule file given a list of descriptors.

Change-Id: If6a148637a88c368c3d9024da5c9d1b50a66bb28