Handle additional exception type when converting BSONElement's
Newer versions of the mongo client changed which type of exception
they throw on BSONElement conversion failures. In previous versions
a mongo::UserException was thrown; now a bson::assertion is thrown.

Conveniently, they also forgot to update their documentation which
clearly states:

    These functions, which start with a capital letter, throw a
    UserException if the element is not of the required type.

So now we catch both mongo::UserException and bson::assertion
exceptions when attempting to convert BSONElement's to arrays.

Change-Id: If6fb8ee41d24b363a12abc647e3bc929f6b697e8