Fix invalid molecule refs with substructure queries
This fixes a bug in which after a substructure query the table
view would fail to lookup molecule refs for molecules after the
user double-clicked on them.

The problem is that the BSONObj stored in each QModelIndex for each
molecule in the main view was not passed through the filter model
which is used for substructure queries. Now, the model index is
properly mapped back to the source model when a filter model is
active on the table.

This also refactors how the molecule details dialog is requested
and displayed when the user clicks either "Show Details" on the
right-click menu or double-clicks a molecule.

Change-Id: I83b162a722efc8370f7bf705a6600005a265e124