Allow for quoted items in CSV files
This adds support for quoted items in CSV files. Certain data
values (such as InChI formulas) contain commas which will break
the CSV parsing. This now allows for the user to group items in
quotes so that the entire value will be parsed as one item instead
of mistakenly splitting it up.

To illustrate, consider parsing the following before and after this
change to the code:

  inchi,name,boiling_point,melting_point
  "InChI=1S/C2H6O/c1-2-3/h3H,2H2",ethanol,78.37,-114

Now, the inchi formula will be properly parsed.

Change-Id: I83f666de0d13acdbdde88da6ccee8d7ab656f202