Improve the CSV Import Dialog
This adds improvements to the CSV import dialog. Now, the user
can set up the mapping between the columns in the CSV file and
the identifiers or descriptors that they should represent in the
database. This also allows for the user use different identifier
types (InChI, InChIKey, SMILES, Name, and CAS) instead of just
InChIKey.

Change-Id: Ib68737c5650d2f715c44113b2a02e40ccd0e9e7d