Add support for rendering SVG diagrams
This adds support for rendering SVG molecular diagrams in the
molecule details dialog. Now, if the molecule document in the
Mongo database contains an 'svg' field the data is loaded and
displayed instead of using the PNG diagram. This allows the
diagram to scale with the size of the dialog and looks better.

This adds a dependance on QtWebKit to perform the SVG rendering.
The QtSvg module was not used because it does not support scaled
rendering and has been deprecated.

Change-Id: I260d4c4b78e542ccf0c28ec0ce629b0743d00041