Updates for the new layout in quantum collections
The computational results model always issues an empty query, which
resulted in the construction fetching all QM documents in the
collections! This easily led to memory exhaustion on big collections,
and was very slow. After that found made some small changes to account
for the updates layout, and to query on the molecule id which serves
as the primary link to the molecule collection.

Change-Id: I0883b4423ad1b2e69d4e4d4f7d96a860e17b4189